Configuração de timezone, locale e padronização dos arquivos de properties

+++ vagrant/shell/install.sh
@@ -1,5 +1,22 @@
 #!/bin/bash
 
-export DEBIAN_FRONTEND=noninteractive # http://serverfault.com/a/670688
+if [ ! -f ~/runonce ]
+then
 
-apt-get install -yq openjdk-8-jdk openjfx
+  export DEBIAN_FRONTEND=noninteractive # http://serverfault.com/a/670688
+
+  # OpenJDK + OpenJFXX
+  apt-get install -yq openjdk-8-jdk openjfx
+
+  # Timezone
+  timedatectl set-timezone America/Sao_Paulo
+
+  # Locale
+  sudo locale-gen pt_BR
+  sudo locale-gen pt_BR.UTF-8
+  sudo dpkg-reconfigure locales
+  sudo update-locale LANG=pt_BR.UTF-8
+
+  touch ~/runonce
+
+fi
